creating separate application timeline component
